+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.9 (WIP)
+## 0.1.9 (2022-07-06)
 
 * Allow supplying a custom `data-src` to `alchemy-field` elements
 * Add `view` templates for File & Datetime fields